% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{.id_type_to_keytype}
\alias{.id_type_to_keytype}
\title{Translate a \code{--id-type} value to an AnnotationDbi keytype}
\usage{
.id_type_to_keytype(id_type)
}
\arguments{
\item{id_type}{One of "ensembl", "entrez", "symbol" (case-insensitive).}
}
\value{
The corresponding AnnotationDbi keytype string.
}
\description{
Internal switch from the user-facing \code{id_type} strings
("ensembl", "entrez", "symbol") to the AnnotationDbi keytype names
("ENSEMBL", "ENTREZID", "SYMBOL").
}
\keyword{internal}
