% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{.msigdbr_species}
\alias{.msigdbr_species}
\title{Map the --annotation flag value to an msigdbr species string}
\usage{
.msigdbr_species(annotation)
}
\arguments{
\item{annotation}{One of "human" or "mouse".}
}
\value{
msigdbr species string ("Homo sapiens" or "Mus musculus").
}
\description{
Internal switch from "human"/"mouse" to the binomial species
names that \code{msigdbr::msigdbr} expects.
}
\keyword{internal}
